Bookmarking Stories to Read Later (#74)
With this diff, I'm taking a pass at bookmarking functionality. This
required adding some infrastructure, namely
[Room](https://developer.android.com/training/data-storage/room) as our
local database.

The action is wired up to a long press for now, which will allow you to
bookmark / unbookmark a story. There is also a separate screen and
bottom nav item for bookmarks so we can easily navigate there.