Updating Previews for Snapshot Baseline (#75)
This PR is just updating and organizing our Previews for Screens and
Components. The ideas is that every component has multiple previews for
different visual states and also different logical states.

Part of the Goal is is to create a `MultiPreview` annotation that we can
just tack on to all future previews with all the desired test states.