Initial Login Infrastructure
This PR adds one core dependency and a bunch of infrastrucutre to support login
and login-based actions.

Because HN doesn't have official API support for Login, Upvotes, Commenting, etc, the only
way that clients can perform those actions is by requesting various web routes and scraping
the HTML for data.

Added `jsoup` as a dependency to make working with HTML a lot nicer, as it also allows you to query HTML documents using CSS selectors.

Also created a `WebClient` which is reponsible for all things that require scraping the Hacker News website.

Lastly I `CookieJar` to our OkHttp client so that it can store the "auth cookie" when we successfully login. We need this cooking on all subsequent web client requests in order to get access to actions.