Fixing a crash when loading Feeds (#82)
Apparently there is a chance that the API just straight up returns null
as a valid response. In this scenario our JsonDeserializer fails and
crashes the app. And in the case that this value is in the first page of
the feed, we will constantly crash on startup.

In order to safeguard against this, I just created a wrapper type and a
custom deserializer that just checks if a gauranteed property exists in
the json response, and if it does, use the normal item deserializer,
otherwise map it to an empty object that is mostly a marker type, we can
just ignore it.