Simple Feed Error Handling
In order to make sure that app doesn't just crash constantly, we need to have some basic error handling in place.

The goal of this diff is to have a very simple solution for most cases. For the feed, we can show an error item that allows you to try request the feed again.

I'd love to have a bit more sophistication here, but this is a good first step. At the very least if they lose internet connection the app won't crash.