Adding About Info to Settings
We discussed adding some attribution and links to an About section. I just added social links and a link to the emerge tools landing page.

I also added a placeholder for a feedback form, which we can fill in once we have a notion page or similar thing setup.

Lastly I wanted to add some fun + emerge-y branding so I created a card that has a sort of starfield animation. I built this a lil while ago and it felt fitting, can remove it if it's too extra.