Comment Hiding + HTML Comments (#108)
This PR adds the ability to hide comment "subtrees".

Because of the way we parse, each comment is an indpendent entity with no parent-child relationship. However, we can use the comment order + indent level to determine if a comment is a child of some arbitrary parent comment. I setup a simple function to find all the children of a comment that would be hidden or displayed based on the selected parent. I also added a lil icon to show show if a comment is displayed or hidden.

Last thing I snuck in here was getting the comment HTML rather than the escaped string, so that we can parse comments as HTML (render links, formatting, etc).