Preview Cleanup + Seperating Components (#120)
This PR seperates out feature components from the screen file. It's
actually a bit cumbersome to leverage previews when there are a BUNCH of
previews in the same file. Also discoverability suffers, so I moved a
bunch of things out into their own component files. This also makes it
easy to see the various states of each component.

It also isolates the screen files to only have previews at the screen
level.