Adding Paging for Feeds (#254)
I built a simple system similar to Android to support paging. We get the
full list of Ids for a feed, and just feed it into a lil helper struct
that will chunk it into pages we can request. Once a page is loaded
fully we return the page of items.

The mechanism for paging is a "marker view" that triggers the viewmodel
to load the next page.