Adding Bookmarking (#305)
This diff adds the infrastructure for bookmarking. It's leveraging
SwiftData, albeit not in the exact way that's recommended since we are
using ViewModels for data retrieval and mapping to state.

There are still probably some kinks to iron out here. I'm calling a lot
of fetches to get the latest version of out bookmarks when ideally I can
just observe some changes.

---------

Co-authored-by: Trevor Elkins <trevor@emergetools.com>